Implement OnReceiveMessage and Max priority logic

+++ internal/network/message.go
@@ -41,7 +41,7 @@ type MsgPropose struct {
 
 type MsgAgree struct {
 	MsgId          string
-	AgreedPriority int
+	AgreedPriority float64
 	FinalNodeID    string
 }
